backend/requirements.txt
fastapi
uvicorn
scipy
numpy
python-dotenv
pyjwt
python-multipart